requirements.txt
opensearch-py==1.0.0
python-dotenv==0.19.2
